major changes to breakout3 because checking out commits results in merge conflicts
Ideally the simplest way to undo these changes is to checkout a past commit, but getting
each attendee to do this results in merge conflicts. Because we don't want to get into
rebasing, we're having attendees just revert their pull requests. This worksheet has them
doing the following:
1. finding their pull request on github
2. reverting it and merging to master
3. viewing repository network graph to visualize branches
4. read some future considerations about issues/project boards, and an alternative way
to undo changes in a repository.